+++ pakyow-support/spec/unit/deprecator/reporters/warn_spec.rb
@@ -1,0 +1,13 @@
+require "pakyow/support/deprecator/reporters/warn"
+
+RSpec.describe Pakyow::Support::Deprecator::Reporters::Warn do
+  describe "::report" do
+    it "warns" do
+      expect(described_class).to receive(:warn).with("[deprecation] message")
+
+      described_class.report {
+        "message"
+      }
+    end
+  end
+end